feat(security): Replace random password generation with user_id and secret key combination
In this commit, a feature is added to the security system. The previous method of generating random passwords is replaced with a new method that generates a password based on a combination of the user's ID and the Django secret key. By using a combination of the user's ID and the Django secret key, the user can be reliably authorized without having to replace existing password each time.