feat(logging): log appwite SDK method call exceptions when debug is enabled
This commit adds logging functionality to the package to log error messages that occur when calling methods in the appwite SDK. The logging is only enabled when the debug flag is turned on, and helps to improve the debugging process by providing a record of errors that occur in package. The logged errors can be used to diagnose and fix issues, improving the stability and reliability of package.