refactor(header): replace usage of USER_ID header with Account object ID
This commit refactors the package to replace the usage of the USER_ID header with the ID received from the Account object. Previously, the USER_ID header was used for generation of password, but now it will be replaced with the ID retrieved from the Account object, ensuring consistency and accuracy.